# wc-guest-management.api/swagger.json
{
	"basePath": "/",
	"consumes": [
		"application/json"
	],
	"definitions": {
		"DashboardMetric": {
			"properties": {
				"slot": {
					"type": "string",
					"enum": [
						"9 AM",
						"12 NN",
						"3 PM",
						"6 PM",
						"NA"
					]
				},
				"count": {
					"type": "number",
					"format": "double"
				}
			},
			"required": [
				"slot",
				"count"
			],
			"type": "object",
			"additionalProperties": true
		},
		"DashboardLineItem": {
			"properties": {
				"label": {
					"type": "string"
				},
				"metrics": {
					"items": {
						"$ref": "#/definitions/DashboardMetric"
					},
					"type": "array"
				}
			},
			"required": [
				"label",
				"metrics"
			],
			"type": "object",
			"additionalProperties": true
		},
		"DashboardCategory": {
			"properties": {
				"title": {
					"type": "string",
					"enum": [
						"age",
						"activity"
					]
				},
				"metrics": {
					"items": {
						"$ref": "#/definitions/DashboardLineItem"
					},
					"type": "array"
				}
			},
			"required": [
				"title",
				"metrics"
			],
			"type": "object",
			"additionalProperties": true
		},
		"DashboardReport": {
			"properties": {
				"totalGuests": {
					"type": "number",
					"format": "double"
				},
				"summary": {
					"items": {
						"$ref": "#/definitions/DashboardMetric"
					},
					"type": "array"
				},
				"categories": {
					"items": {
						"$ref": "#/definitions/DashboardCategory"
					},
					"type": "array"
				}
			},
			"required": [
				"totalGuests",
				"summary",
				"categories"
			],
			"type": "object",
			"additionalProperties": true
		},
		"Guest": {
			"properties": {
				"_id": {
					"type": "string",
					"x-nullable": true
				},
				"visitDate": {
					"type": "string",
					"format": "date-time"
				},
				"tableNumber": {
					"type": "number",
					"format": "double"
				},
				"volunteer": {
					"type": "string"
				},
				"guest": {
					"type": "string"
				},
				"age": {
					"type": "number",
					"format": "double",
					"x-nullable": true
				},
				"birthDate": {
					"type": "string",
					"x-nullable": true
				},
				"mobile": {
					"type": "string",
					"x-nullable": true
				},
				"email": {
					"type": "string",
					"x-nullable": true
				},
				"civilStatus": {
					"type": "string",
					"x-nullable": true
				},
				"cityOfResidence": {
					"type": "string",
					"x-nullable": true
				},
				"cityOfWorkplace": {
					"type": "string",
					"x-nullable": true
				},
				"category": {
					"type": "string",
					"x-nullable": true
				},
				"series": {
					"type": "number",
					"format": "double",
					"x-nullable": true
				},
				"createdDate": {
					"type": "string",
					"format": "date-time",
					"x-nullable": true
				},
				"worshipDay": {
					"type": "string",
					"x-nullable": true
				},
				"worshipTime": {
					"type": "string",
					"enum": [
						"9 AM",
						"12 NN",
						"3 PM",
						"6 PM",
						"NA"
					],
					"x-nullable": true
				},
				"action": {
					"type": "string",
					"enum": [
						"A",
						"DNA",
						"Prayed",
						"Counseled"
					],
					"x-nullable": true
				},
				"gender": {
					"type": "string",
					"x-nullable": true
				}
			},
			"required": [
				"visitDate",
				"tableNumber",
				"volunteer",
				"guest"
			],
			"type": "object",
			"additionalProperties": true
		},
		"InfoSlip": {
			"properties": {
				"visitDate": {
					"type": "string"
				},
				"worshipTime": {
					"type": "string",
					"enum": [
						"9 AM",
						"12 NN",
						"3 PM",
						"6 PM",
						"NA"
					],
					"x-nullable": true
				},
				"tableNumber": {
					"type": "number",
					"format": "double",
					"x-nullable": true
				},
				"volunteer": {
					"type": "string"
				},
				"guests": {
					"type": "string"
				}
			},
			"required": [
				"visitDate",
				"volunteer",
				"guests"
			],
			"type": "object",
			"additionalProperties": true
		}
	},
	"info": {
		"title": "wc-guest-management-api",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		}
	},
	"paths": {
		"/api/dashboard": {
			"get": {
				"operationId": "GetDashboardReport",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"$ref": "#/definitions/DashboardReport"
						}
					}
				},
				"security": [
					{
						"bearer": []
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "fromDate",
						"required": false,
						"format": "date-time",
						"type": "string"
					},
					{
						"in": "query",
						"name": "toDate",
						"required": false,
						"format": "date-time",
						"type": "string"
					}
				]
			}
		},
		"/api/guest": {
			"get": {
				"operationId": "FetchGuests",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"items": {
								"$ref": "#/definitions/Guest"
							},
							"type": "array"
						}
					}
				},
				"security": [
					{
						"bearer": []
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "fromDate",
						"required": false,
						"format": "date-time",
						"type": "string"
					},
					{
						"in": "query",
						"name": "toDate",
						"required": false,
						"format": "date-time",
						"type": "string"
					},
					{
						"in": "query",
						"name": "criteria",
						"required": false,
						"type": "string"
					},
					{
						"in": "query",
						"name": "slot",
						"required": false,
						"type": "string",
						"enum": [
							"9 AM",
							"12 NN",
							"3 PM",
							"6 PM",
							"NA"
						]
					}
				]
			}
		},
		"/api/guest/{id}": {
			"put": {
				"operationId": "UpdateGuestData",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"$ref": "#/definitions/Guest"
						}
					}
				},
				"security": [
					{
						"bearer": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"type": "string"
					},
					{
						"in": "body",
						"name": "body",
						"schema": {
							"properties": {
								"guestData": {
									"$ref": "#/definitions/Guest"
								}
							},
							"title": "UpdateGuestDataBody",
							"type": "object",
							"required": [
								"guestData"
							]
						}
					}
				]
			}
		},
		"/api/guest/category/{category}": {
			"get": {
				"operationId": "FetchGuestsByCategory",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"items": {
								"$ref": "#/definitions/Guest"
							},
							"type": "array"
						}
					}
				},
				"security": [
					{
						"bearer": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "category",
						"required": true,
						"type": "string",
						"enum": [
							"age",
							"activity"
						]
					},
					{
						"in": "query",
						"name": "slot",
						"required": false,
						"type": "string",
						"enum": [
							"9 AM",
							"12 NN",
							"3 PM",
							"6 PM",
							"NA"
						]
					},
					{
						"in": "query",
						"name": "index",
						"required": false,
						"type": "string"
					},
					{
						"in": "query",
						"name": "fromDate",
						"required": false,
						"format": "date-time",
						"type": "string"
					},
					{
						"in": "query",
						"name": "toDate",
						"required": false,
						"format": "date-time",
						"type": "string"
					}
				]
			}
		},
		"/api/guest/welcome": {
			"post": {
				"operationId": "Welcome",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"items": {
								"$ref": "#/definitions/Guest"
							},
							"type": "array"
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "body",
						"name": "body",
						"schema": {
							"properties": {
								"infoSlip": {
									"$ref": "#/definitions/InfoSlip"
								},
								"print": {
									"type": "boolean"
								}
							},
							"title": "WelcomeBody",
							"type": "object",
							"required": [
								"infoSlip"
							]
						}
					}
				]
			}
		},
		"/api/profile/auth": {
			"post": {
				"operationId": "Auth",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"type": "string"
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "body",
						"name": "body",
						"schema": {
							"properties": {
								"email": {
									"type": "string"
								},
								"password": {
									"type": "string"
								}
							},
							"title": "AuthBody",
							"type": "object",
							"required": [
								"email",
								"password"
							]
						}
					}
				]
			}
		},
		"/api/profile/me": {
			"get": {
				"operationId": "Me",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"additionalProperties": true,
							"type": "object"
						}
					}
				},
				"security": [
					{
						"bearer": []
					}
				],
				"parameters": []
			}
		}
	},
	"produces": [
		"application/json"
	],
	"swagger": "2.0",
	"securityDefinitions": {
		"bearer": {
			"type": "apiKey",
			"in": "header",
			"name": "authorization"
		}
	}
}